#886 Use the desktop download limit if not on a metered connection

+++ octoprint/src/main/java/de/crysxd/octoapp/octoprint/models/settings/Settings.kt
@@ -43,8 +43,8 @@ data class Settings(
     interface PluginSettings
 
     data class GcodeViewerSettings(
-        val mobileSizeThreshold: Int,
-        val sizeThreshold: Int
+        val mobileSizeThreshold: Long,
+        val sizeThreshold: Long
     ) : PluginSettings
 
     data class TradfriSettings(
